Add copy and delete gist functionality
Enhances buttons on gist view to copy and delete gists

+++ lib/draconic_systems/gists.ex
@@ -4,6 +4,7 @@ defmodule DraconicSystems.Gists do
   """
 
   import Ecto.Query, warn: false
+  alias DraconicSystems.Accounts.User
   alias DraconicSystems.Repo
 
   alias DraconicSystems.Gists.Gist
@@ -86,8 +87,15 @@ defmodule DraconicSystems.Gists do
       {:error, %Ecto.Changeset{}}
 
   """
-  def delete_gist(%Gist{} = gist) do
-    Repo.delete(gist)
+  def delete_gist(%User{} = user, gist_id) do
+    gist = Repo.get(Gist, gist_id)
+
+    if user.id == gist.user_id do
+      Repo.delete(gist)
+      {:ok, gist}
+    else
+      {:error, :unauthorized}
+    end
   end
 
   @doc """
